Bug fix in sentence level
